modified:   lovelypay-paper/src/main/java/org/lovelysmp/lovelypay/config/types/MainConfig.java 	modified:   lovelypay-paper/src/main/java/org/lovelysmp/lovelypay/config/types/MocNapConfig.java 	modified:   lovelypay-paper/src/main/java/org/lovelysmp/lovelypay/config/types/MocNapServerConfig.java 	modified:   lovelypay-paper/src/main/java/org/lovelysmp/lovelypay/handler/banking/payos/PayosHandler.java 	modified:   lovelypay-paper/src/main/java/org/lovelysmp/lovelypay/handler/banking/payos/data/PayosPayment.java 	modified:   lovelypay-paper/src/main/java/org/lovelysmp/lovelypay/service/MilestoneService.java

+++ lovelypay-paper/src/main/java/org/lovelysmp/lovelypay/handler/banking/payos/data/PayosPayment.java
@@ -6,10 +6,10 @@ import lombok.Builder;
 @Builder
 public class PayosPayment {
     @SerializedName("orderCode")
-    private int orderCode;
+    private long orderCode;
 
     @SerializedName("amount")
-    private double amount;
+    private long amount;
 
     @SerializedName("description")
     private String description;
